requirements.txt
jupyterlab
requests
tiingo
iexfinance
PyYAML
jupyter_contrib_nbextensions
jupyter_nbextensions_configurator
plotly
ipywidgets
dash
bokeh
jupyterlab-dash
numpy
pandas